Rename package money.boku.transfers to just money.boku
