+++ src/main/java/money/boku/Main.java
@@ -1,7 +1,8 @@
-package money.boku.transfers;
+package money.boku;
 
 import io.javalin.Javalin;
 import io.javalin.http.HttpStatus;
+import money.boku.account.*;
 import money.boku.transfers.account.*;
 
 /**
