+++ src/main/java/money/boku/MoneyService.java
@@ -1,7 +1,7 @@
-package money.boku.transfers;
+package money.boku;
 
-import money.boku.transfers.account.AccountNotFoundException;
-import money.boku.transfers.account.NegativeBalanceRequestException;
+import money.boku.account.AccountNotFoundException;
+import money.boku.account.NegativeBalanceRequestException;
 
 import java.math.BigDecimal;
 import java.util.UUID;
